Added Spawn Points with spawn point selection based on player distance
